Refactor auth route and add token middleware
