+++ server/middleware/verifyToken.js
@@ -1,0 +1,15 @@
+const jwt = require("jsonwebtoken");
+
+function verifyToken(req, res, next) {
+  const token = req.headers.authorization?.split(" ")[1];
+  if (!token) return res.sendStatus(401);
+  try {
+    const decoded = jwt.verify(token, process.env.JWT_SECRET || "SECRET_KEY");
+    req.user = decoded;
+    next();
+  } catch (err) {
+    res.sendStatus(403);
+  }
+}
+
+module.exports = verifyToken;
