+++ server/routes/posts.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const db = require("../db");
+const verifyToken = require("../middleware/verifyToken");
 
 router.get("/", (req, res) => {
   db.query("SELECT * FROM posts ORDER BY id DESC", (err, results) => {
@@ -16,7 +17,7 @@ router.get("/:id", (req, res) => {
   });
 });
 
-router.post("/", (req, res) => {
+router.post("/", verifyToken, (req, res) => {
   const { title, content } = req.body;
   db.query("INSERT INTO posts (title, content) VALUES (?, ?)", [title, content], (err, result) => {
     if (err) return res.status(500).send(err);
@@ -24,14 +25,14 @@ router.post("/", (req, res) => {
   });
 });
 
-router.delete("/:id", (req, res) => {
+router.delete("/:id", verifyToken, (req, res) => {
   db.query("DELETE FROM posts WHERE id = ?", [req.params.id], (err) => {
     if (err) return res.status(500).send(err);
     res.sendStatus(200);
   });
 });
 
-router.put("/:id", (req, res) => {
+router.put("/:id", verifyToken, (req, res) => {
   const { title, content } = req.body;
   db.query("UPDATE posts SET title=?, content=? WHERE id=?", [title, content, req.params.id], (err) => {
     if (err) return res.status(500).send(err);
